MeteoMeter/Ionic/src/pages/login/forgotPassword/login-forgotPassword.spec.ts

import { IonicModule, NavController, Platform } from 'ionic-angular';
import { NavMock } from '../../../../test-config/mocks/navMock';
import { PlatformMock } from '../../../../test-config/mocks/platformMock';
import { SplashScreen } from '@ionic-native/splash-screen';
import { SplashScreenMock } from '../../../../test-config/mocks/splashScreenMock';
import { StatusBar } from '@ionic-native/status-bar';
import { StatusBarMock } from '../../../../test-config/mocks/statusBarMock';
import { Title } from '@angular/platform-browser';
import { TitleMock } from '../../../../test-config/mocks/titleMock';
import { async, ComponentFixture, fakeAsync, flushMicrotasks, TestBed } from '@angular/core/testing';
import { TranslateModule, TranslateService } from '@ngx-translate/core';
import { TranslateServiceMock } from '../../../../test-config/mocks/translateServiceMock';
import { LanguagePageModule } from '../../language/language.module';
import { LanguagePage } from '../../language/language';
import { MeteoMeter } from '../../../app/app.component';
import { LoginForgotPassword } from "./login-forgotPassword";
import { UserAAAAA } from '../../../services/userAAAAA.service';
import { UserAAAAAMock } from '../../../../test-config/mocks/userAAAAAMock';
describe('LoginForgotPassword Component', () => {
  let fixture: ComponentFixture<LoginForgotPassword>;
  let component: LoginForgotPassword;
  let spy: jasmine.Spy;

  beforeEach(async(() => {
    TestBed.configureTestingModule({
      declarations: [
        MeteoMeter,
        LoginForgotPassword
      ],
      imports: [
        LanguagePageModule,
        TranslateModule.forChild(),
        IonicModule.forRoot(MeteoMeter)
      ],
      providers: [
        { provide: UserAAAAA, useClass: UserAAAAAMock },
        { provide: Platform, useClass: PlatformMock },
        { provide: StatusBar, useClass: StatusBarMock },
        { provide: SplashScreen, useClass: SplashScreenMock },
        { provide: TranslateService, useClass: TranslateServiceMock },
        { provide: NavController, useClass: NavMock },
        { provide: Title, useClass: TitleMock },
      ]
    }).compileComponents();
  }));

  beforeEach(async() => {
    fixture = TestBed.createComponent(LoginForgotPassword);
    component = fixture.componentInstance;
  });
  
  afterEach(() => {
    fixture.destroy();
    component = null;
  });

  it('LoginForgotPassword Component: should be created', fakeAsync(() => {
    expect(component instanceof LoginForgotPassword).toBeTruthy();
  }));

  it('LoginForgotPassword Component: should navigate to language page', fakeAsync(() => {
    spy = spyOn(component, 'onLanguage').and.callThrough();
    component.onLanguage();
    flushMicrotasks();
    expect(spy).toHaveBeenCalled();
    spy = spyOn(component._nav, 'push').and.callThrough();
    component.onLanguage();
    flushMicrotasks();
    expect(spy).toHaveBeenCalledWith(LanguagePage);
    spy.calls.mostRecent().returnValue.then((result) => {
      expect(result).toEqual('PUSH');
    });
  }));
  it('LoginForgotPassword Component - onCheckEmail: should use userAAAAA service to CheckEmail', fakeAsync(() => {
    var email = 'test';
    spy = spyOn(component, 'onCheckEmail').and.callThrough();
    component.onCheckEmail(
    email
    );
    flushMicrotasks();
    expect(spy).toHaveBeenCalledWith(
    email
    );
    spy = spyOn(component._userAAAAA, 'checkEmail').and.callThrough();
    component.onCheckEmail(
    email
    );
    flushMicrotasks();
    expect(spy).toHaveBeenCalledWith(
    email
    );
    var result = spy.calls.mostRecent().returnValue;
    expect(result).toBeDefined();
  }));
  it('LoginForgotPassword Component - onToLogin: should navigate to login-main page', fakeAsync(() => {
    spy = spyOn(component, 'onToLogin').and.callThrough();
    component.onToLogin();
    flushMicrotasks();
    expect(spy).toHaveBeenCalled();
    spy = spyOn(component._nav, 'push').and.callThrough();
    component.onToLogin();
    flushMicrotasks();
    expect(spy).toHaveBeenCalledWith('login-main');
    spy.calls.mostRecent().returnValue.then((result) => {
      expect(result).toEqual('PUSH');
    });
  }));

  it('LoginForgotPassword Component - onSendMail: should do nothing', fakeAsync(() => {
    spy = spyOn(component, 'onSendMail').and.callThrough();
    component.onSendMail();
    flushMicrotasks();
    expect(spy).toHaveBeenCalled();
    var result = spy.calls.mostRecent().returnValue;
    expect(result).toBeUndefined();
  }));

});
